refactor: 2차 피드백 반영
