+++ README.md
@@ -1,5 +1,8 @@
 ## MVP, UI 테스트
 
+### 5단계 - 리팩터링
+- [x] 2차 피드백 반영
+
 ### 4단계 - UI 테스트
 - [x] 1차 피드백 반영
 - [x] 입력된 피연산자가 없는 경우 숫자 버튼 누를 경우 해당 숫자가 화면에 표시
